// src/app/core/shared/models/basket-model.ts
export interface Field {
  visibility: boolean;
  validators: Array<string | { name: string }>;
}